backend/requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.27.0
Flask-JWT-Extended==4.3.1
flask-cors==3.0.10
Flask-Mail==0.9.1
python-dateutil==2.8.1
requests==2.27.1
gunicorn==20.1.0